+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 - Added support for CQL filter to ``WFS``, credits to `Eilio <https://github.com/emiliom>`__.
 - Moved all the web services URLs to a YAML file that ``ServiceURL`` class reads. It makes
   managing the new URLs easier. The file is located at ``pygeoogc/static/urls.yml``.
+- Turned off threading by default for all the services since not all web services supports it.
 
 0.1.10 (2020-08-18)
 -------------------